requirements.txt
scikit-learn>=0.19.1
pandas>=0.22.0
nltk>=3.2.5
numpy
scipy
pytorch